src/main/kotlin/com/multhink/http/security/AuthenticationClient.kt
/*
 * Copyright (c) 2023 Krystian Piotr Kryszczak.
 * All rights reserved.
 */
package com.multhink.http.security

import io.micronaut.http.annotation.Body
import io.micronaut.http.annotation.Post
import io.micronaut.http.client.annotation.Client
import io.micronaut.security.authentication.UsernamePasswordCredentials
import io.micronaut.security.token.render.BearerAccessRefreshToken

@Client("/login")
interface AuthenticationClient {
    @Post
    fun login(@Body credentials: UsernamePasswordCredentials): BearerAccessRefreshToken
}
